# Optimize bitmask finding some more.