```
| Before [e26fb39e] <main>   | After [9753dfea] <optimize-again>   |   Ratio | Benchmark (Parameter)                                      |
|----------------------------|-------------------------------------|---------|------------------------------------------------------------|
| 4.10±0.05ms                | 3.64±0.04ms                         |    0.89 | cohorts.ERA5MonthHour.time_find_group_cohorts              |
| 4.56±0.05ms                | 4.00±0.06ms                         |    0.88 | cohorts.ERA5MonthHourRechunked.time_find_group_cohorts     |
| 3.31±0.08ms                | 2.79±0.01ms                         |    0.84 | cohorts.ERA5DayOfYear.time_find_group_cohorts              |
| 8.15±0.05ms                | 6.78±0.02ms                         |    0.83 | cohorts.OISST.time_find_group_cohorts                      |
| 659±10μs                   | 519±6μs                             |    0.79 | cohorts.PerfectBlockwiseResampling.time_find_group_cohorts |
| 663±20μs                   | 487±5μs                             |    0.73 | cohorts.PerfectMonthly.time_find_group_cohorts             |
| 2.75±0.03ms                | 1.93±0.03ms                         |    0.7  | cohorts.ERA5Google.time_find_group_cohorts                 |
```